Move console-related logic to new project
