+++ ConsoleExample/Components/Example/Events/AlarmEvent.cs
@@ -1,5 +1,5 @@
 ﻿using Summer.Events.Interfaces;
 
-namespace Summer.Components.Example.Events;
+namespace ConsoleExample.Components.Example.Events;
 
 public record AlarmEvent(ClockTime AlarmTime) : IEvent;
