+++ ConsoleExample/Components/Example/Events/TimeChangedEvent.cs
@@ -1,0 +1,6 @@
+﻿
+using Summer.Events.Interfaces;
+
+namespace ConsoleExample.Components.Example.Events;
+
+public record TimeChangedEvent(ClockTime Time) : IEvent;
