+++ Summer.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Summer", "Summer\Summer.csp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SummerTests", "SummerTests\SummerTests.csproj", "{BC74AA30-63A8-41F7-BC5C-AFCA6EFF3717}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleExample", "ConsoleExample\ConsoleExample.csproj", "{DF29BAED-BF83-4844-9F1B-6A78DB2FEC40}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{BC74AA30-63A8-41F7-BC5C-AFCA6EFF3717}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BC74AA30-63A8-41F7-BC5C-AFCA6EFF3717}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BC74AA30-63A8-41F7-BC5C-AFCA6EFF3717}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DF29BAED-BF83-4844-9F1B-6A78DB2FEC40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DF29BAED-BF83-4844-9F1B-6A78DB2FEC40}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DF29BAED-BF83-4844-9F1B-6A78DB2FEC40}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DF29BAED-BF83-4844-9F1B-6A78DB2FEC40}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
