--- Summer/Components/Example/Events/TimeChangedEvent.cs
@@ -1,5 +1,0 @@
-﻿using Summer.Events.Interfaces;
-
-namespace Summer.Components.Example.Events;
-
-public record TimeChangedEvent(ClockTime Time) : IEvent;
